Remove code duplication by introducing AbstractRestProblemIT class
