+++ belgif-rest-problem-it/belgif-rest-problem-it-common/pom.xml
@@ -32,6 +32,11 @@
       <artifactId>assertj-core</artifactId>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>io.rest-assured</groupId>
+      <artifactId>rest-assured</artifactId>
+      <scope>provided</scope>
+    </dependency>
   </dependencies>
 
 </project>
