+++ belgif-rest-problem-it/belgif-rest-problem-spring-it/src/test/java/io/github/belgif/rest/problem/RestProblemSpringIT.java
@@ -1,11 +1,9 @@
 package io.github.belgif.rest.problem;
 
-import static org.hamcrest.Matchers.*;
+import java.util.Arrays;
+import java.util.stream.Stream;
 
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
+import org.junit.jupiter.api.Disabled;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 
@@ -13,82 +11,25 @@ import io.restassured.RestAssured;
 import io.restassured.specification.RequestSpecification;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class RestProblemSpringIT {
+class RestProblemSpringIT extends AbstractRestProblemIT {
 
     @LocalServerPort
     private int port;
 
-    private RequestSpecification spec;
-
-    @BeforeEach
-    void before() {
-        spec = RestAssured.with().baseUri("http://localhost").port(port).basePath("/spring/frontend");
+    @Override
+    protected RequestSpecification getSpec() {
+        return RestAssured.with().baseUri("http://localhost").port(port).basePath("/spring/frontend");
     }
 
-    @Test
-    void badRequest() {
-        spec.when().get("/badRequest").then().assertThat()
-                .statusCode(400)
-                .body("type", equalTo("urn:problem-type:belgif:badRequest"))
-                .body("detail", equalTo("Bad Request from frontend"));
+    @Override
+    protected Stream<String> getClients() {
+        return Arrays.stream(Client.values()).map(Client::name);
     }
 
-    @Test
-    void custom() {
-        spec.when().get("/custom").then().assertThat()
-                .statusCode(409)
-                .body("type", equalTo("urn:problem-type:acme:custom"))
-                .body("customField", equalTo("value from frontend"));
-    }
-
-    @Test
-    void runtime() {
-        spec.when().get("/runtime").then().assertThat()
-                .statusCode(500)
-                .body("type", equalTo("urn:problem-type:belgif:internalServerError"));
-    }
-
-    @Test
-    void unmapped() {
-        spec.when().get("/unmapped").then().assertThat()
-                .statusCode(400)
-                .body("type", equalTo("urn:problem-type:belgif:test:unmapped"))
-                .body("detail", equalTo("Unmapped problem from frontend"));
-    }
-
-    @Test
-    void retryAfter() {
-        spec.when().get("/retryAfter").then().assertThat()
-                .statusCode(503)
-                .header("Retry-After", "10000")
-                .body("type", equalTo("urn:problem-type:belgif:serviceUnavailable"));
-    }
-
-    @ParameterizedTest
-    @EnumSource
-    void badRequestFromBackend(Client client) {
-        spec.when().get("/badRequestFromBackend?client=" + client).then().assertThat()
-                .statusCode(400)
-                .body("type", equalTo("urn:problem-type:belgif:badRequest"))
-                .body("detail", equalTo("Bad Request from backend (caught successfully by frontend)"));
-    }
-
-    @ParameterizedTest
-    @EnumSource
-    void customFromBackend(Client client) {
-        spec.when().get("/customFromBackend?client=" + client).then().assertThat()
-                .statusCode(409)
-                .body("type", equalTo("urn:problem-type:acme:custom"))
-                .body("customField", equalTo("value from backend (caught successfully by frontend)"));
-    }
-
-    @ParameterizedTest
-    @EnumSource
-    void unmappedFromBackend(Client client) {
-        spec.when().get("/unmappedFromBackend?client=" + client).then().assertThat()
-                .statusCode(400)
-                .body("type", equalTo("urn:problem-type:belgif:test:unmapped"))
-                .body("detail", equalTo("Unmapped problem from backend (caught successfully by frontend)"));
+    @Override
+    @Disabled("Not supported yet: https://github.com/belgif/rest-problem-java/issues/12")
+    void beanValidation() {
+        super.beanValidation();
     }
 
 }
